[REFACTOR/#275] 약관/온보딩 패키지 원복
