+++ app/src/main/res/layout/activity_sign_up.xml
@@ -16,7 +16,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        android:name="com.example.teumteum.ui.signup.AgreementFragment"
+        android:name="com.example.teumteum.ui.onboarding.AgreementFragment"
         />
 
     <!-- 상단 진행바 -->
