+++ app/src/main/res/layout/fragment_agreement.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.signup.AgreementFragment">
+    tools:context=".ui.onboarding.AgreementFragment">
 
     <!-- 제목 -->
     <TextView
